changing fields to int from float

+++ vocode/streaming/models/transcriber.py
@@ -64,8 +64,8 @@ class TranscriberConfig(TypedModel, type=TranscriberType.BASE.value):
     input_device_config: Optional[InputDeviceConfig] = None
     vad: bool = False
     experimental: bool = False
-    utterance_end_ms: float = None
-    deepgram_endpoint_ms: float = None
+    utterance_end_ms: int = None
+    deepgram_endpoint_ms: int = None
 
     @validator("min_interrupt_confidence")
     def min_interrupt_confidence_must_be_between_0_and_1(cls, v):
